// apps/api/src/modules/accounts/users/enums/user-sort-by.enum.ts
// src/modules/accounts/users/enums/user-sort-by.enum.ts
export enum UserSortBy {
  CREATED_AT = 'createdAt',
  UPDATED_AT = 'updatedAt',
  USERNAME = 'username',
  FIRST_NAME = 'firstName',
  LAST_NAME = 'lastName',
  EMAIL = 'email',
  ROLE = 'role',
  PROVIDER = 'provider',
  IS_ACTIVE = 'isActive',
  LAST_LOGIN_AT = 'lastLoginAt',
}